+++ example/lib/app/controller_app.dart
@@ -5,4 +5,4 @@ export 'controller/example_app_controller.dart';
 export 'logging/controller/log_controller.dart';
 
 //
-export 'settings/_controller_settings.dart';
+export 'settings/controller_settings.dart';