+++ example/lib/app/view_app.dart
@@ -2,12 +2,7 @@
 export 'logging/view/log_page.dart';
 
 //
-export 'view/my_app.dart';
-
-//
-//export 'two_tabs_scaffold.dart';
-//
-export 'settings/_view_settings.dart';
+export 'settings/view_settings.dart';
 
 //
 export 'view/common/dev_settings.dart';
@@ -20,6 +15,9 @@ export 'view/common/splash_screen.dart';
 
 //
 export 'view/multi_tabs/multi_tabs_scaffold.dart';
+
+//
+export 'view/my_app.dart';
 
 //
 export 'view/object/_view_app_object.dart';