+++ example/lib/src/controller.dart
@@ -2,6 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 //
+
 export 'package:state_extended/state_extended.dart' show StateXController;
 
 //