+++ lib/part09_builder_statefulwidget.dart
@@ -6,8 +6,8 @@ part of 'state_extended.dart';
 
 // Call the AppState's builder() function
 class _BuilderStatefulWidget extends StatefulWidget {
-  const _BuilderStatefulWidget(this.appState);
-  final AppStateX appState;
+  const _BuilderStatefulWidget();
+
   @override
   State<StatefulWidget> createState() => _BuilderState();
 }
@@ -18,14 +18,21 @@ class _BuilderState extends State<_BuilderStatefulWidget> {
   @override
   void initState() {
     super.initState();
-    appState = widget.appState;
-    appState._builderState = this; // nullified in AppStateX dispose()
+    appState = context.findAncestorStateOfType<AppStateX>();
+    appState?._builderState = this; // nullified in AppStateX dispose()
   }
 
   // Will never be null
-  late AppStateX appState;
+  AppStateX? appState;
 
   //
   @override
-  Widget build(BuildContext context) => appState.builder(context);
+  Widget build(BuildContext context) =>
+      appState?.builder(context) ?? const SizedBox.shrink();
+
+  @override
+  void dispose() {
+    appState = null;
+    super.dispose();
+  }
 }